commit x

+++ 0x14-bit_manipulation/100-get_endianness.c
@@ -10,5 +10,5 @@ int get_endianness(void)
 {
 	unsigned int i;
 
-	return ((int) (((char *)&i)[0]));
+	return ((((char *)&i)[0]));
 }
